✨ feat(config): persist the sidebar orientation in .gwm.toml
`[tui] sidebar_position` survived a restart; the orientation did not.
`SidebarState::orientation` was runtime-only, so a user who prefers
side-by-side had to press the cycle key on every single launch.

Move `SidebarOrientation` from `tui::state::sidebar` to `config`, where
`SidebarPosition` / `MacroOpenMode` / `TuiOpenMode` already live, derive
serde on it, and add `[tui] sidebar_orientation`. The state module
re-exports the enum so existing paths keep resolving.

`rename_all = "kebab-case"`, not `lowercase`: the latter spells
`SideBySide` as `sidebyside` — the trap `MacroOpenMode` hit on PR #292.
It also keeps the serialised form equal to `label()`, so the Settings
panel write-back produces a file that still loads.

Seeded at both sites `sidebar_position` uses (construction + config
reload) and exposed in the Settings TUI tab next to `sidebar position`.

The default stays `stacked` (#217): `#[serde(default)]` resolves through
the enum's existing `#[default]`. The `orientation` doc-comment claimed
the default was `auto` — stale since #217, and exactly the kind of thing
that talks someone into seeding the wrong value. Corrected here.

refs #365
